Release 160214
Updated bake template

+++ src/Template/Bake/Element/Controller/delete.ctp
@@ -21,14 +21,16 @@
      * @return void Redirects to index.
      * @throws \Cake\Network\Exception\NotFoundException When record not found.
      */
-    public function delete($id = null)
-    {
-        $this->request->allowMethod(['post', 'delete']);
+    public function delete($id = null) {
         $<%= $singularName %> = $this-><%= $currentModelName %>->get($id);
-        if ($this-><%= $currentModelName; %>->delete($<%= $singularName %>)) {
-            $this->Flash->success('The <%= strtolower($singularHumanName) %> has been deleted.');
-        } else {
-            $this->Flash->error('The <%= strtolower($singularHumanName) %> could not be deleted. Please, try again.');
+        if ($this->request->is(['patch', 'post', 'put'])) {
+            if ($this-><%= $currentModelName; %>->delete($<%= $singularName %>)) {
+                $this->Flash->success('The <%= strtolower($singularHumanName) %> has been deleted.');
+            } else {
+                $this->Flash->error('The <%= strtolower($singularHumanName) %> could not be deleted. Please, try again.');
+            }
+            return $this->redirect(['action' => 'index']);
         }
-        return $this->redirect(['action' => 'index']);
+        $this->set(compact('<%= $singularName %>'));
     }
+
